# perf(stream): significantly improve stream lock performance