--- a/Cargo.toml
+++ b/Cargo.toml
@@ -46,11 +46,14 @@ tokio-util = { version = "0.7.1", features = ["codec", "io"] }
 tokio = { version = "1", features = ["io-util"] }
 bytes = "1"
 http = "1"
-tracing = { version = "0.1", default-features = false, features = ["std"], optional = true }
+tracing = { version = "0.1", default-features = false, features = [
+    "std",
+], optional = true }
 fnv = "1.0.5"
 slab = "0.4.2"
 indexmap = { version = "2", features = ["std"] }
 smallvec = "1.15.0"
+parking_lot = "0.12.4"
 
 [dev-dependencies]
 tracing = { version = "0.1", default-features = false, features = ["std"] }
@@ -66,7 +69,12 @@ serde = "1.0.0"
 serde_json = "1.0.0"
 
 # Examples
-tokio = { version = "1", features = ["rt-multi-thread", "macros", "sync", "net"] }
+tokio = { version = "1", features = [
+    "rt-multi-thread",
+    "macros",
+    "sync",
+    "net",
+] }
 env_logger = { version = "0.10", default-features = false }
 
 tokio-rustls = "0.26"