## Benchmark

1. request num: 100000

```log
===============================
Starting single-threaded server at 127.0.0.1:5928
-------------------------------
Single-threaded client: h2_send_requests at 127.0.0.1:5928
Overall: 288ms.
Fastest: 199ms
Slowest: 263ms
Avg    : 224ms
-------------------------------
Single-threaded client: parking_lot_h2_send_requests at 127.0.0.1:5928
Overall: 299ms.
Fastest: 183ms
Slowest: 281ms
Avg    : 236ms
-------------------------------
Multi-threaded client: h2_send_requests at 127.0.0.1:5928
Overall: 246ms.
Fastest: 1ms
Slowest: 384ms
Avg    : 243ms
-------------------------------
Multi-threaded client: parking_lot_h2_send_requests at 127.0.0.1:5928
Overall: 9ms.
Fastest: 0ms
Slowest: 52ms
Avg    : 23ms
===============================
Starting multi-threaded server at 127.0.0.1:5929
-------------------------------
Single-threaded client: h2_send_requests at 127.0.0.1:5929
Overall: 450ms.
Fastest: 89ms
Slowest: 429ms
Avg    : 367ms
-------------------------------
Single-threaded client: parking_lot_h2_send_requests at 127.0.0.1:5929
Overall: 423ms.
Fastest: 85ms
Slowest: 403ms
Avg    : 354ms
-------------------------------
Multi-threaded client: h2_send_requests at 127.0.0.1:5929
Overall: 346ms.
Fastest: 1ms
Slowest: 489ms
Avg    : 335ms
-------------------------------
Multi-threaded client: parking_lot_h2_send_requests at 127.0.0.1:5929
Overall: 125ms.
Fastest: 0ms
Slowest: 239ms
Avg    : 137ms
```

3. request num: 1000000

```log
===============================
Starting single-threaded server at 127.0.0.1:5928
-------------------------------
Single-threaded client: h2_send_requests at 127.0.0.1:5928
Overall: 3106ms.
Fastest: 1838ms
Slowest: 2968ms
Avg    : 2433ms
-------------------------------
Single-threaded client: parking_lot_h2_send_requests at 127.0.0.1:5928
Overall: 3065ms.
Fastest: 1827ms
Slowest: 2938ms
Avg    : 2408ms
-------------------------------
Multi-threaded client: h2_send_requests at 127.0.0.1:5928
Overall: 3546ms.
Fastest: 32ms
Slowest: 4864ms
Avg    : 3492ms
-------------------------------
Multi-threaded client: parking_lot_h2_send_requests at 127.0.0.1:5928
Overall: 423ms.
Fastest: 17ms
Slowest: 1265ms
Avg    : 488ms
===============================
Starting multi-threaded server at 127.0.0.1:5929
-------------------------------
Single-threaded client: h2_send_requests at 127.0.0.1:5929
Overall: 4496ms.
Fastest: 598ms
Slowest: 4358ms
Avg    : 3779ms
-------------------------------
Single-threaded client: parking_lot_h2_send_requests at 127.0.0.1:5929
Overall: 4393ms.
Fastest: 609ms
Slowest: 4265ms
Avg    : 3711ms
-------------------------------
Multi-threaded client: h2_send_requests at 127.0.0.1:5929
Overall: 4033ms.
Fastest: 1ms
Slowest: 6457ms
Avg    : 4013ms
-------------------------------
Multi-threaded client: parking_lot_h2_send_requests at 127.0.0.1:5929
Overall: 2924ms.
Fastest: 17ms
Slowest: 4411ms
Avg    : 2986ms
```